second commit

+++ modules/voila3sample-be/src/main/java/it/micegroup/voila3sample/service/PrivilegeService.java
@@ -8,47 +8,43 @@ import org.springframework.data.jpa.domain.Specification;
 
 import java.util.Optional;
 
-public interface PrivilegeService extends BaseEntityService<Privilege, Integer> {
+public interface PrivilegeService extends BaseEntityService<Privilege, Long> {
 
-	/**
-	 * Print Pdf and Xlxs
-	 */
-	public byte[] printPdfReport(String objectKey);
+  /** Print Pdf and Xlxs */
+  public byte[] printPdfReport(String objectKey);
 
-	byte[] printXLSList(Specification<Privilege> specification);
+  byte[] printXLSList(Specification<Privilege> specification);
 
-	/**
-	 * Executes the bulk update of a Privilege
-	 * 
-	 * @param privilege
-	 * @return updated Privilege
-	 */
-	Privilege bulkUpdate(Privilege privilege);
+  /**
+   * Executes the bulk update of a Privilege
+   *
+   * @param privilege
+   * @return updated Privilege
+   */
+  Privilege bulkUpdate(Privilege privilege);
 
-	/**
-	 * Executes the update of an Privilege
-	 * 
-	 * @param privilege
-	 * @return Privilege updated
-	 */
-	Privilege update(Privilege privilege);
+  /**
+   * Executes the update of an Privilege
+   *
+   * @param privilege
+   * @return Privilege updated
+   */
+  Privilege update(Privilege privilege);
 
-	/**
-	 * Deletes an Privilege with the objectKey in input, if it exists
-	 * 
-	 * @param objectKey
-	 * @return Privilege deleted if it was present
-	 */
-	Optional<Privilege> delete(String objectKey);
+  /**
+   * Deletes an Privilege with the objectKey in input, if it exists
+   *
+   * @param objectKey
+   * @return Privilege deleted if it was present
+   */
+  Optional<Privilege> delete(String objectKey);
 
-	/**
-	 * Returns the Page of the Privilege following the specification in input
-	 * 
-	 * @param specification
-	 * @param pageable
-	 * 
-	 * @return Page of Privilege
-	 */
-	Page<Privilege> search(Specification<Privilege> specification, Pageable pageable);
-
+  /**
+   * Returns the Page of the Privilege following the specification in input
+   *
+   * @param specification
+   * @param pageable
+   * @return Page of Privilege
+   */
+  Page<Privilege> search(Specification<Privilege> specification, Pageable pageable);
 }
